LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-ja.html

<!DOCTYPE html><!-- webkit-test-runner [ language=ja ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "ja");
</script>
</body>
</html>
